Move process_for_categories
