+++ data-workflows/plugin/metadata.py
@@ -2,10 +2,10 @@ import logging
 from collections import defaultdict
 from typing import Optional
 
-from nhcommons.utils.categories import process_for_categories
 from nhcommons.utils.custom_parser import render_description
 from nhcommons.utils.github_adapter import get_github_metadata, is_valid_repo_url
 from nhcommons.utils.pypi_adapter import get_plugin_pypi_metadata
+from plugin.categories import process_for_categories
 
 logger = logging.getLogger(__name__)
